+++ ChangeLog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+## 1.3.9 - 2021.03.10
+
 - Feature: Improve stability for general and fluorescent markers gridding process.
 - Feature: Support for general marker chip (e.g. ZION (Z1C), Kenai (K1B)) gridding process.
 - Feature: Parameterize the statistics window size ratio for evluating the probe intensity whose coefficient of variation is the smallest.